denoise debug logs and take suggestions
